Move API view classes under common/public
